requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
pydantic==2.11.3
python-dotenv==1.1.0
requests==2.32.3
openai==0.27.0
python-multipart==0.0.5
